requirements.txt
weaviate_client==2.5.0
openpyxl==3.0.5
PyYAML==5.4.1
